move globals templates
