+++ src/pages/[[...slug]].tsx
@@ -8,10 +8,10 @@ import dynamic from 'next/dynamic'
 import Script from 'next/script'
 import { drupalClient } from '@/lib/drupal/drupalClient'
 import { getGlobalElements } from '@/lib/drupal/getGlobalElements'
-import { Wrapper } from '@/templates/globals/wrapper'
+import { Wrapper } from '@/templates/layouts/wrapper'
 import { NewsStory } from '@/templates/layouts/newsStory'
 import { StoryListing } from '@/templates/layouts/storyListing'
-import HTMLComment from '@/templates/globals/util/HTMLComment'
+import HTMLComment from '@/templates/common/util/HTMLComment'
 import { shouldHideHomeBreadcrumb } from '@/lib/utils/breadcrumbs'
 import { Event } from '@/templates/layouts/event'
 import { EventListing } from '@/templates/layouts/eventListing'
@@ -23,12 +23,12 @@ import {
 } from '@/lib/drupal/staticProps'
 import { StaticPropsResource } from '@/lib/drupal/staticProps'
 import { FormattedResource } from '@/data/queries'
-import { LayoutProps } from '@/templates/globals/wrapper'
+import { LayoutProps } from '@/templates/layouts/wrapper'
 import { NewsStory as FormattedNewsStory } from '@/types/formatted/newsStory'
 import { StoryListing as FormattedStoryListing } from '@/types/formatted/storyListing'
 import { EventListing as FormattedEventListing } from '@/types/formatted/eventListing'
 import { Event as FormattedEvent } from '@/types/formatted/event'
-import { Meta } from '@/templates/globals/meta'
+import { Meta } from '@/templates/common/meta'
 import { PreviewCrumb } from '@/templates/common/preview'
 
 const RESOURCE_TYPES_TO_BUILD = [
