+++ src/templates/common/banners/facilityBanner/index.tsx
