+++ src/templates/common/banners/promoBanner/index.tsx
