+++ src/templates/common/banners/promoBanner/promoBanner.stories.mdx
