+++ src/templates/common/footer/index.tsx
