+++ src/templates/common/medallia/index.tsx
