+++ src/templates/layouts/wrapper/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
-import { Banner } from '@/templates/globals/banners/banner'
-import { PromoBanner } from '@/templates/globals/banners/promoBanner'
-import { FacilityBanner } from '@/templates/globals/banners/facilityBanner'
+import { Banner } from '@/templates/common/banners/banner'
+import { PromoBanner } from '@/templates/common/banners/promoBanner'
+import { FacilityBanner } from '@/templates/common/banners/facilityBanner'
 import { HeaderFooterData } from '@/types/formatted/headerFooter'
 import { BannersData } from '@/types/formatted/banners'
 import { NodeBannerType } from '@/types/drupal/node'
@@ -9,8 +9,8 @@ import { handleSkipLink } from '@/lib/utils/handleSkipLink'
 import { UnpublishedBanner } from '@/templates/common/preview'
 import { StaticPropsResource } from '@/lib/drupal/staticProps'
 import { FormattedResource } from '@/data/queries'
-import { Footer } from '../footer'
-import { Header } from '../header'
+import { Footer } from '../../common/footer'
+import { Header } from '../../common/header'
 
 // Allows additions to window object without overwriting global type
 interface customWindow extends Window {
